requirements.txt
SpeechRecognition
pygame
pyautogui
keyboard
pyaudio
vosk
numpy
newsapi-python
gtts
pyperclip
requests
beautifulsoup4
wikipedia
pyjokes
yt-dlp
geocoder
psutil
pypiwin32
comtypes
pyttsx3
spacy
nltk
scikit-learn
# Optional / advanced
openai-whisper
torch
transformers
